backend/BillardApp/testsite/requirements/base.txt
Django==2.0.3
mysqlclient==1.3.12
pytz==2018.3
djangorestframework==3.7.7
